requirements.txt
construct>=2.9.45
lxml>=4.2.3
pycryptodome>=3.6.4
requests>=2.19.1